+++ PHPUnit/TextUI/TestRunner.php
@@ -276,6 +276,10 @@ class PHPUnit_TextUI_TestRunner extends PHPUnit_Runner_BaseTestRunner
               $arguments['addUncoveredFilesFromWhitelist']
             );
 
+            $codeCoverage->setCheckForUnintentionallyCoveredCode(
+              $arguments['strict']
+            );
+
             $codeCoverage->setProcessUncoveredFilesFromWhitelist(
               $arguments['processUncoveredFilesFromWhitelist']
             );